refactor: isolate optimizer effect analysis
